IANA considerations: complete registration info
This is modeled after draft-irtf-cfrg-chacha20-poly1305-10 § 5 (the RFC
Editor changes the IANA Consideration sections after publication).
The registry is defined in RFC 5116 § 6 and defined an entry to consist
of a short name, a positive number and a reference to a specification.
The request now also contains the reference.

This commit deviates from the naming used in the aforementioned draft
because the formal name in RFC 5116 § 6 is "AEAD Registry".