Use <sup></sup> to indicate superscript
This enhances rendering in all formats except plaintext, where it still
is largely equivalent.
The plaintext rendering leads to an unfortunate collision with
the ^ (XOR) operator defined in the document, but readers can probably
be expected to realize the different contexts.

The XML tag is passed through with no modification to xml2rfc, which
handles it correctly.

There is no native kramdown-rfc2629 syntax for this.