Reorder sections highest-level first
It feels odd to start with the Update() function.
While in C function declaration order, Update() would indeed have to
come first -- barring forward declarations --, this is an unnatural
starting point for reading and building outside in.
The benefits of lessening cognitive load while reading outweigh the
downsides of not being immediately compatible with a C-based
implementation.