Address review concerns in #12.
Don't reference "[c]urrently standardized AEAD schemes" because this
phrasing would not age well.
Without taking into account the age of the document, such a statement
may quickly become obsolete.
Remove the mention of AES-GCM and ChaCha20-Poly1305 at the same time
while at it;
it is not entirely inconceivable that a later RFC amends the IETF's
definitions of AES-GCM and ChaCha20-Poly1305 without renaming them.
This part may still be clarified during e.g. a mailing list discussion
on the CFRG.

For similar reasons, remove mention of a concrete speed-up comparison to
AES-GCM.
While it is undeniable that the algorithm per se is faster,
the concrete value of course differs between platforms and
implementations.