Encoding (ct, tag) tuples: require order
This really should be a MUST because if a protocol incorporates this
construction, the format should be at least consistent across
implementations.
We still allow deviation in the form of keeping the tag in a separate
structure.

The order specified is consistent with RFC 8439 (ChaPoly) § 2.8,
so if we require this, it's just the same as it's always been.