Security Considerations: miscellaneous
This was originally intended (as a continuation of #21) to establish
compliance with BCP 72.
Upon closer inspection, most things that need saying were already said;
in particular, the threat model of an AEAD scheme is generally implied
by virtue of being an AEAD scheme (cf. RFC 8439 § 4).
The only change of this nature here is that the security properties of
AEGIS break if you reuse the nonce (i.e. nonce reuse attacks are out of
scope).

The rest of the changes are:

- instances of applying BCP 14 keywords that were missed in
b7c34444024f8275c7c319cbcd2aceafef450377; and
- reordering to make the most general statements start at the top
in hopes of achieving more logical ordering.