Wipe msg before on error in Decrypt()
As per feedback from Daniel Bleichenbacher (meeting 2022-03-09),
it is vital that we ensure no plaintext is released.
While large steps towards this have already been made with the
changes in -03 already,
this makes it abundantly clear what to do,
especially in languages like C,
where you might indeed leave the buffer alone and just return the error
(perhaps forgetting about the partial message);
we mandate by spec that it must be wiped instead.