Change msg = {} to a note after the text
Introduced in 06e700a93aef898cc1d5ea5fc179f0b2ad82fd30,
review by Frank Denis in #39 noted that the meaning of msg = {} is
unclear.
Instead introduce a description of what to do.

If the function is faithfully implemented 1:1,
then there is no issue as the newly-created plaintext message is never
returned.
However, if the function works by modifying a user-supplied buffer
(which is a common pattern in C in particular),
it is easy to forget wiping the message on failure.
Therefore, this specific case should be called out as implementation
guidance.