AEGIS-X: include the parallelism degree in the context
This is a breaking change, and it is not strictly necessary, but since
it is not too late, it may be desirable.

Currently, if a (key,nonce) is reused with different variants of the
same security level, the key streams initialized with the same context
will be the same.

For example, with AEGIS-128L and AEGIS-128X2, the 128L key stream will
be the same as the first half of the 128X2 key stream.

Tags will be different, but reusing a (key,nonce) with different variants
would still reveal the difference between the messages.

We can document that (key,nonce) pairs must not be reused with different
algorithms.

But a better alternative would be to include the degree in the context.

We actually use (degree - 1), so that when the degree is 1, the context
remains all-zero, and it doesn't break the equivalence between AEGIS-128L
and AEGIS-128X1.