AEGISMAC-128L and 256: add the tag length, as a context separator
AEGISMAC-128X and AEGISMAC-256X have the nice property that
128-bit tags cannot be computed from 256-bit tags, and that
the dedicated MAC functions don't output the same tags as
the encryption functions.

The 128L and 256 variants didn't had that context separation.

But this is trivial and free to do, by setting msg_len to
the tag length.

Swap the context-separation halves of MAC-128X and MAC-256X
for consistency.