Consistently use bits, even for the tag length
In FinalizeMac, there was an inconsistency between the input
length and the tag length. The former being in bits, the latter
in bytes.

While this is fine, it can be confusing. Especially when
implementing the 128L and 256 versions as 128X1 and 256X1.

Switch to bits everywhere instead of mixing units.